Move tests to different function
